Added Club windup. Changed wealth change code to work with decimals instead of doubles to fix weird rounding error. Removed active spells when recast a second time by the same spellcaster. Added GetWeaponTarget function with support for filtering. Added ReplaceDamageDice function. Added Weapon Effects for the Quarterstaff and the Club. Fleshed out the Shillelagh spell. Added support for changing the weapon's magic hue from the spell logic (with OverrideWeaponMagicHue). Added HandsOnWeapon property to PlayerActionShortcut to show how many hands were on a weapon. Added Game_PickWeapon event handler to select a player's weapon.

+++ DndCore/Core Classes/Life/Creature.cs
@@ -66,7 +66,7 @@ namespace DndCore
 		public ObservableCollection<ItemViewModel> equipment = new ObservableCollection<ItemViewModel>();
 
 		public double flyingSpeed = 0;
-		public double goldPieces = 0;
+		public decimal goldPieces = 0;
 
 		public double hitPoints = 0;
 		public double initiative = 0;
